+++ app/javascript/flavours/glitch/features/compose/util/counter.js
